requirements.txt
# Core
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.6
sqlalchemy==2.0.9
alembic==1.10.3

# JSON Processing
ijson==3.2.3
orjson==3.8.12

# Language Model
google-generativeai==0.3.2

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0

# Frontend
streamlit==1.35.0
requests==2.31.0
openai==1.3.3
